[DOCKER] Allow building with local source instead of Git - This fixes issues with extending projects to build `master` by default unless they take it into account per branch. This is easily forgotten and this fix also gets rid of the dependency with Git. We want to make it as easy as possible to build a database, not introduce more complexity. - If the `DBSOURCE_PATH` set exists on the filesystem when trying to build it will be used as the source code needed to build the database. Every Git specific validation/action will not be done in such a case. - This is also used to build our own Docker image. So no Git download needed there either. - Added a proper wait for PostgreSQL to exit cleanly. This probably didn't cause any issues as the shell script will only exit when all child processes are done, but still. This makes it explicit. The nerd in me doesn't mind the explicitness. - Building from root of the project now. This makes it easier for us to add the source without the need for a copy. Moved and updated the `update.sh` and `build_images.sh` script to reflect this. This also gets rid of the issue of copying `build-database.sh` to every Docker image version directory. As it's not a template, this is preferred. Auto build via Docker Hub might still need this, but we'll cross that bridge when we get to it. - Added a bit more logging.

+++ docker/build-database.sh
@@ -1,14 +1,9 @@
-#!/bin/sh
+#!/bin/bash
 
 set -e
 
 # validations
 : ${SFTP_READONLY_PASSWORD?'SFTP_READONLY_PASSWORD must be provided'}
-: ${GIT_USERNAME?'GIT_USERNAME must be provided'}
-: ${GIT_TOKEN?'GIT_TOKEN must be provided'}
-: ${GIT_HOSTNAME?'GIT_HOSTNAME must be provided'}
-: ${GIT_ORG?'GIT_ORG must be provided'}
-: ${GIT_REPOSITORY?'GIT_REPOSITORY must be provided'}
 : ${DBDATA_PATH?'DBDATA_PATH must be provided'}
 : ${DBSOURCE_PATH?'DBSOURCE_PATH must be provided'}
 : ${DATABASE_NAME?'DATABASE_NAME must be provided'}
@@ -17,50 +12,87 @@ set -e
 : ${DBCONFIG_PATH:=aerius-database-build-config/config}
 : ${DBRUNSCRIPT:=default}
 
-# add git dependencies
-apk --no-cache add --virtual .git-deps git openssh
+# If the source code isn't made available by the extending Dockerfile we need to use git to check it out
+USE_GIT=! [[ -d "${DBSOURCE_PATH}" ]]
+
+[[ ${USE_GIT} ]] && echo 'Source code present in GIT..'
+! [[ ${USE_GIT} ]] && echo 'Source code present locally..'
+
+# Do git specific validations and fix paths if needed
+if [[ ${USE_GIT} ]]; then
+  : ${GIT_USERNAME?'GIT_USERNAME must be provided'}
+  : ${GIT_TOKEN?'GIT_TOKEN must be provided'}
+  : ${GIT_HOSTNAME?'GIT_HOSTNAME must be provided'}
+  : ${GIT_ORG?'GIT_ORG must be provided'}
+  : ${GIT_REPOSITORY?'GIT_REPOSITORY must be provided'}
+  # Default to githash if there is no version explicitly set
+  : ${DATABASE_VERSION:=#}
+
+  DBSOURCE_PATH="${GIT_REPOSITORY}/${DBSOURCE_PATH}"
+  DBCONFIG_PATH="${GIT_REPOSITORY}/${DBCONFIG_PATH}"
+else
+  # We require a version outside of git
+   : ${DATABASE_VERSION?'DATABASE_VERSION must be set if source code is present locally..'}
+fi
+
+# add git dependencies if needed
+[[ ${USE_GIT} ]] && apk --no-cache add --virtual .git-deps git openssh
 
 # Make our own PGDATA.
 # We are unfortunately doing this because we want the data to persist but the default PGDATA directory is marked as a volume, which cannot be undone.
 mkdir -p "${PGDATA}" && chown -R postgres:postgres "${PGDATA}" && chmod 777 "${PGDATA}"
 
-# fetch repo
-git --version
-git clone "https://${GIT_USERNAME}:${GIT_TOKEN}@${GIT_HOSTNAME}/${GIT_ORG}/${GIT_REPOSITORY}.git"
+# fetch repo if needed
+[[ ${USE_GIT} ]] && git --version
+[[ ${USE_GIT} ]] && git clone "https://${GIT_USERNAME}:${GIT_TOKEN}@${GIT_HOSTNAME}/${GIT_ORG}/${GIT_REPOSITORY}.git"
 
 # create db-data folder for the repo
 mkdir -p "${DBDATA_PATH}"
 
 # configure repo with the readonly SFTP password given as argument
-echo "\$sftp_data_readonly_password = '${SFTP_READONLY_PASSWORD}'" > "${GIT_REPOSITORY}/${DBCONFIG_PATH}/AeriusSettings.User.rb"
+echo "\$sftp_data_readonly_password = '${SFTP_READONLY_PASSWORD}'" >> "${DBCONFIG_PATH}/AeriusSettings.User.rb"
+
+# Set git support off in the build script
+! [[ ${USE_GIT} ]] && echo "\$vcs = :none" >> "${DBCONFIG_PATH}/AeriusSettings.User.rb"
 
 # sync db-data files we need
-cd "${GIT_REPOSITORY}/${DBSOURCE_PATH}/"
-ruby ../../aerius-database-build/bin/SyncDBData.rb settings.rb --from-sftp --to-local
+echo 'Syncing database data files..'
+cd "${DBSOURCE_PATH}/"
+ruby /aerius-database-build/bin/SyncDBData.rb settings.rb --from-sftp --to-local
 
 # initialize database
 # (this is a wrapper provided by the postgres image, which will initialize the db if it isn't already and is executed when starting the image for the first time.
-#  Run 'postgres' which will trigger the initialization and start the database, so we can start building the database.)
+# Run 'postgres' which will trigger the initialization and start the database, so we can start building the database.)
 /usr/local/bin/docker-entrypoint.sh postgres &
 
 # The database starts twice. First to set it up and a second time to simply start cleanly.
 # Wait for it to stop once. (Detect socket file being removed)
+echo 'Waiting for PostgreSQL to stop..'
 inotifywait -e DELETE --include .s.PGSQL.5432 /var/run/postgresql/
 
 # Wait for the DB to finish starting up (the second time)
+echo 'Waiting for PostgreSQL to start up again..'
 while [[ ! -S /var/run/postgresql/.s.PGSQL.5432 ]]; do
   sleep 0.5s
 done
 
 # execute database build
-ruby ../../aerius-database-build/bin/Build.rb "${DBRUNSCRIPT}" settings.rb -v '#' -n "${DATABASE_NAME}"
+ruby /aerius-database-build/bin/Build.rb "${DBRUNSCRIPT}" settings.rb -v "${DATABASE_VERSION}" -n "${DATABASE_NAME}"
 
 # make the image smaller by doing a VACUUM FULL ANALYZE
 psql -U "${POSTGRES_USER}" -d "${DATABASE_NAME}" -c 'VACUUM FULL ANALYZE'
 
-# stop PostgreSQL database cleanly
+# stop PostgreSQL database cleanly by sending a kill signal
 kill %1
 
+# Wait for the PostgreSQL process to end
+wait %1
+
 # image cleanup (removing unneeded db-data, git directory and git dependencies)
-rm -rf /"${DBDATA_PATH}" "${GIT_REPOSITORY}"
-apk del .git-deps
+rm -rf "${DBDATA_PATH}"
+[[ ${USE_GIT} ]] && rm -rf "${GIT_REPOSITORY}"
+[[ ${USE_GIT} ]] && apk del .git-deps
+
+# Exit with 0 if this stage is reached, otherwise the return code from
+#  the last if statement might be used, which might let Docker think the build failed
+exit 0
